+++ boto3/resources/factory.py
@@ -105,7 +105,8 @@ class ResourceFactory(object):
 
         # Attributes that get auto-loaded
         self._load_attributes(
-            attrs=attrs, meta=meta, resource_model=resource_model,
+            attrs=attrs, meta=meta, resource_name=resource_name,
+            resource_model=resource_model,
             service_context=service_context)
 
         # Collections and their corresponding methods
@@ -168,7 +169,8 @@ class ResourceFactory(object):
                 action_model=action, resource_name=resource_name,
                 service_context=service_context)
 
-    def _load_attributes(self, attrs, meta, resource_model, service_context):
+    def _load_attributes(self, attrs, meta, resource_name, resource_model,
+                         service_context):
         """
         Load resource attributes based on the resource shape. The shape
         name is referenced in the resource JSON, but the shape itself
@@ -187,10 +189,18 @@ class ResourceFactory(object):
         for name, (orig_name, member) in attributes.items():
             if name in identifiers:
                 prop = self._create_identifier_alias(
-                    identifier=identifiers[name], member_model=member)
+                    resource_name=resource_name,
+                    identifier=identifiers[name],
+                    member_model=member,
+                    service_context=service_context
+                )
             else:
                 prop = self._create_autoload_property(
-                    name=orig_name, snake_cased=name, member_model=member)
+                    resource_name=resource_name,
+                    name=orig_name, snake_cased=name,
+                    member_model=member,
+                    service_context=service_context
+                )
             attrs[name] = prop
 
     def _load_collections(self, attrs, resource_model, service_context):
@@ -274,7 +284,8 @@ class ResourceFactory(object):
 
         return property(get_identifier)
 
-    def _create_identifier_alias(factory_self, identifier, member_model):
+    def _create_identifier_alias(factory_self, resource_name, identifier,
+                                 member_model, service_context):
         """
         Creates a read-only property that aliases an identifier.
         """
@@ -283,15 +294,18 @@ class ResourceFactory(object):
 
         get_identifier.__name__ = str(identifier.member_name)
         get_identifier.__doc__ = docstring.AttributeDocstring(
+            service_name=service_context.service_name,
+            resource_name=resource_name,
             attr_name=identifier.member_name,
+            event_emitter=factory_self._emitter,
             attr_model=member_model,
             include_signature=False
         )
 
         return property(get_identifier)
 
-    def _create_autoload_property(factory_self, name, snake_cased,
-                                  member_model):
+    def _create_autoload_property(factory_self, resource_name, name,
+                                  snake_cased, member_model, service_context):
         """
         Creates a new property on the resource to lazy-load its value
         via the resource's ``load`` method (if it exists).
@@ -312,7 +326,10 @@ class ResourceFactory(object):
 
         property_loader.__name__ = str(snake_cased)
         property_loader.__doc__ = docstring.AttributeDocstring(
+            service_name=service_context.service_name,
+            resource_name=resource_name,
             attr_name=snake_cased,
+            event_emitter=factory_self._emitter,
             attr_model=member_model,
             include_signature=False
         )