+++ tests/functional/test_session.py
@@ -36,3 +36,13 @@ class TestSession(unittest.TestCase):
     def test_can_access_region_property(self):
         session = boto3.session.Session(region_name='us-west-1')
         self.assertEqual(session.region_name, 'us-west-1')
+
+    def test_get_available_partitions(self):
+        partitions = self.session.get_available_partitions()
+        self.assertIsInstance(partitions, list)
+        self.assertTrue(partitions)
+
+    def test_get_available_regions(self):
+        regions = self.session.get_available_regions('s3')
+        self.assertIsInstance(regions, list)
+        self.assertTrue(regions)