+++ tests/unit/docs/__init__.py
@@ -46,11 +46,17 @@ class BaseDocsTest(unittest.TestCase):
         self.paginator_json_model = {}
         self.resource_json_model = {}
         self._setup_models()
-        self._write_models()
 
         self.doc_name = 'MyDoc'
         self.doc_structure = DocumentStructure(self.doc_name)
 
+        self.setup_client_and_resource()
+
+    def tearDown(self):
+        shutil.rmtree(self.root_dir)
+
+    def setup_client_and_resource(self):
+        self._write_models()
         self.loader = Loader(extra_search_paths=[self.root_dir])
         self.botocore_session = botocore.session.get_session()
         self.botocore_session.register_component('data_loader', self.loader)
@@ -58,9 +64,6 @@ class BaseDocsTest(unittest.TestCase):
             botocore_session=self.botocore_session, region_name='us-east-1')
         self.client = self.session.client('myservice', 'us-east-1')
         self.resource = self.session.resource('myservice', 'us-east-1')
-
-    def tearDown(self):
-        shutil.rmtree(self.root_dir)
 
     def _setup_models(self):
         self.json_model = {